src/requirements.txt
matplotlib==3.7.1
numpy==1.25.0
optuna==3.3.0
pandas==1.5.3
pillow==9.4.0
pyarrow==11.0.0
scikit-learn==1.2.2
scipy==1.10.1
streamlit==1.24.1
streamlit-drawable-canvas==0.9.3
